Refactor: フォルダ配置の整理
