+++ next.config.ts
@@ -5,6 +5,7 @@ import { fileURLToPath } from "url";
 const rootDir = dirname(fileURLToPath(import.meta.url));
 
 const nextConfig: NextConfig = {
+  transpilePackages: ["@kpool/types", "@kpool/wiki"],
   turbopack: {
     root: rootDir,
   },
