+++ packages/wiki/package.json
@@ -1,0 +1,12 @@
+{
+  "name": "@kpool/wiki",
+  "version": "0.0.0",
+  "private": true,
+  "type": "module",
+  "dependencies": {
+    "@kpool/types": "workspace:*"
+  },
+  "exports": {
+    ".": "./src/index.ts"
+  }
+}
