+++ pnpm-workspace.yaml
@@ -1,1 +1,3 @@
 minimumReleaseAge: 20160
+packages:
+  - "packages/*"
